% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{createBinaryGrid}
\alias{createBinaryGrid}
\title{Create binary raster grid of class RasterLayer based on polygon boundary.}
\usage{
createBinaryGrid(polybound, resolution, nrows = NA, ncols = NA, ext = NA,
  vis = F, outFile = "")
}
\arguments{
\item{polybound}{Polygon of class 'SpatialPolygonsDataFrame'}

\item{nrows}{Number of rows. If specified, overrides `res` argument}

\item{ncols}{Number of columns. If specified, overrides `res` argument}

\item{ext}{Alternative to using polybound, specifies extent. Matrix bounding
box format: matrix(c(xmin,ymin,xmax,ymax),2,2)}

\item{vis}{TRUE/FALSE. Do you want a plot made of the outcome? default =  FALSE}

\item{outFile}{Optional. Filename for raster, including extension; if entered,
will write to disk}

\item{res}{Resolution for raster cells. Numeric vector of length 1 or 2.}
}
\description{
This function takes the extent and CRS from boundary polygon, and creates a
raster object with specified cell size where active cells have value = 1 and
inactive cells outside boundary have value = 0. Cell size can be specified by
resolution or nrows and ncols. The extent can be specified if differnt
from the polygon input, which is used to assign active cells.
}
\examples{
# specify resolution  ------------------
poly <- readOGR('S:/Users/deinesji/HPA/gis/lhm', 'RRB_bound2_6500m', verbose=F)
cellSize = 1000
outFileName = 'S:/Users/deinesji/HPA/gis/lhm/rrb1_surfaceGrid_1000m.img'

# create raster object in R
test <- createBinaryGrid(polybound = poly, resolution = cellSize)
# manually view
plot(test)
plot(poly, add=T)

create raster object in R, plot result
test <- createBinaryGrid(polybound = poly, resolution = cellSize, vis = T)

# create raster object in R, plot result, write out Raster
test <- createBinaryGrid(polybound = poly, resolution = cellSize, vis = T,
                        outFile = outFileName)

# specify nrow/ncell  ---------------
poly <- readOGR('S:/Users/deinesji/HPA/gis/lhm', 'RRB_bound2_6500m', verbose=F)
nr = 286
nc = 496
outFileName = 'S:/Users/deinesji/HPA/gis/lhm/rrb1_surfaceGrid_1000m_test.img'

# create raster object in R
test2 <- createBinaryGrid(polybound = poly, ncols = nc, nrows = nr)

# specify extent ------------------
poly <- readOGR('S:/Users/deinesji/HPA/gis/lhm', 'RRB_bound2_6500m', verbose=F)
cellSize = 1000
outFileName = 'S:/Users/deinesji/HPA/gis/lhm/rrb1_surfaceGrid_1000m.img'
# create extent matrix
bbox <- matrix(c(-245000,-153000,273000,154000),2,2)
# create raster object in R
test3 <- createBinaryGrid(polybound = poly, resolution = cellSize, ext = bbox)
}
\keyword{binary}
\keyword{raster}

