% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHeadH5}
\alias{getHeadH5}
\title{Get head from GMS output}
\usage{
getHeadH5(h5file, rTemplate, layer, startYear)
}
\arguments{
\item{h5file}{Filepath to exported .h5 file}

\item{rTemplate}{Model grid raster of Formal class RasterLayer}

\item{layer}{Model layer to extract head. Only layer = 1 works.}

\item{startYear}{Starting year of your model, used to name raster grids}
}
\description{
This function translates head values exported as .h5 from GMS into rasters
}
\examples{
getHeadH5()
}
\keyword{head}

