% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{getHeadObsSim}
\alias{getHeadObsSim}
\title{Extract simulated and observed head values from MODFLOW files}
\usage{
getHeadObsSim(hobfile, osfile, sort = TRUE)
}
\arguments{
\item{hobfile}{Filepath to .hob file}

\item{osfile}{Filepath to ._os file}

\item{sort}{TRUE/FALSE. default =  TRUE. If true, the returned data.frame is
sorted by time, then by well ID.}
}
\description{
This function combines head metadata from the .hob file and head data from
the ._os file, returning a cleaned dataset of obname, wellID, x and y
coordinates, time, and simulated and observed head values.
}
\examples{
# specify model files
fname_hob <- 'BJ_3L_93_AnnualObs.hob'
fname_os <- 'BJ_3L_93_AnnualObs._os'

# run function
simObs <- getHeadObsSim(fname_hob, fname_os, sort = T)
}
\keyword{head}
\keyword{observed}
\keyword{simulated}

