% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groudwater-package.R
\docType{package}
\name{groudwater-package}
\alias{groudwater}
\alias{groudwater-package}
\title{groudwater: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
